@@ -1,18 +1,19 @@
 package com.nikolai.mynews.Controllers.Views;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.bumptech.glide.RequestManager;
 import com.nikolai.mynews.Controllers.Models.TopStoriesArticle;
 import com.nikolai.mynews.R;
+import com.squareup.picasso.Picasso;
 
 import java.util.List;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by Cynthia Nikolai on 7/11/2019.
@@ -20,11 +21,14 @@ import java.util.List;
 
 public class TopStoriesArticleAdapter extends RecyclerView.Adapter<TopStoriesArticleViewHolder> {
 
-    private static final String TAG = "TopStoriesArtAdapter";
+    private static final String TAG = TopStoriesArticleAdapter.class.getSimpleName();
+    private TextView mTextView;
 
     public interface Listener {
-        void onClickDeleteButton(int position);
-        void changeArticleColor(int position, TextView textview, Context context);
+        //void onClickDeleteButton(int position);
+        //void changeArticleColor(int position, TextView textview, Context context);
+        //void addToReadQueue(int position);
+        //boolean hasBeenRead(int position);
     }
 
     // FOR COMMUNICATION
@@ -32,12 +36,13 @@ public class TopStoriesArticleAdapter extends RecyclerView.Adapter<TopStoriesArt
 
     // FOR DATA
     private final List<TopStoriesArticle> mTopStoriesArticles;
-    private final RequestManager glide;
+    //private final List<Integer> mTopStoriesArticlesThatHaveBeenRead = new ArrayList<>();
+    private final Picasso picasso;
 
     // CONSTRUCTOR
-    public TopStoriesArticleAdapter(List<TopStoriesArticle> topStoriesArticles, RequestManager glide, Listener callback) {
+    public TopStoriesArticleAdapter(List<TopStoriesArticle> topStoriesArticles, Picasso picasso, Listener callback) {
         this.mTopStoriesArticles = topStoriesArticles;
-        this.glide = glide;
+        this.picasso = picasso;
         this.callback = callback;
     }
 
@@ -72,9 +77,21 @@ public class TopStoriesArticleAdapter extends RecyclerView.Adapter<TopStoriesArt
         this.mTopStoriesArticles.remove(position);
     }
 
-//    public void changeArticleColor(int position, View textview, Context context){
-//        Log.d(TAG, "changeArticleColor: inside");
-//        textview.setTextColor(ContextCompat.getColor(context, R.color.orange));
+//    public void changeArticleColor(int position, TextView textview, Context context) {
+//        //get id of article to be changed
+//        int id = getPositionID(position);
+//        //add id of article to be changed to already read list
+//        mTopStoriesArticlesThatHaveBeenRead.add(id);
+//        //textview.setTextColor((ContextCompat.getColor(context, R.color.orange)));
 //    }
+
+//    public void addToReadQueue(String id) {
+//        ArticleBeenRead.getInstance().setArticleHasBeenRead(id);
+//    }
+
+//    public int getPositionID(int position){
+//        return this.mTopStoriesArticles.get(position).getId();
+//    }
+
 
 }
