@@ -2,23 +2,23 @@ package com.nikolai.mynews.Controllers.Views;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.RequestOptions;
+import com.nikolai.mynews.Controllers.Fragments.ArticleBeenRead;
 import com.nikolai.mynews.Controllers.Models.TopStoriesArticle;
 import com.nikolai.mynews.Controllers.Models.URL;
 import com.nikolai.mynews.R;
+import com.squareup.picasso.Picasso;
 
 import java.lang.ref.WeakReference;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Locale;
 
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
@@ -28,7 +28,7 @@ import butterknife.ButterKnife;
 
 public class TopStoriesArticleViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
-    private static final String TAG = "TopStoriesArticleViewHolder";
+    private static final String TAG = TopStoriesArticleViewHolder.class.getSimpleName();
 
     private final DateFormat dateFormat = new SimpleDateFormat("MM/dd/yy", Locale.US);
 
@@ -47,13 +47,14 @@ public class TopStoriesArticleViewHolder extends RecyclerView.ViewHolder impleme
         super(itemView);
         this.context=context;
         ButterKnife.bind(this, itemView);
+        itemView.setOnClickListener(this);
     }
 
     public void updateWithNewsArticle(TopStoriesArticle TopStoriesArticle, TopStoriesArticleAdapter.Listener callback) {
-        RequestOptions options = new RequestOptions()
-                .centerCrop()
-                .placeholder(R.mipmap.ic_launcher_round)
-                .error(R.mipmap.ic_launcher_round);
+//        RequestOptions options = new RequestOptions()
+//                .centerCrop()
+//                .placeholder(R.mipmap.ic_launcher_round)
+//                .error(R.mipmap.ic_launcher_round);
         String ending = "";
 
         if (TopStoriesArticle.getMultimedia().size() == 0) {
@@ -63,6 +64,13 @@ public class TopStoriesArticleViewHolder extends RecyclerView.ViewHolder impleme
             ending = " > "+ TopStoriesArticle.getSubsection();
         }
 
+        //check to see if the id is in the read id's or not and change the color
+        if (ArticleBeenRead.getInstance().hasBeenRead(TopStoriesArticle.getUrl())) {
+            this.textView.setTextColor(ContextCompat.getColor(context, R.color.orange));
+        }
+        else {
+            this.textView.setTextColor(ContextCompat.getColor(context, android.R.color.black));
+        }
         this.textView.setText(TopStoriesArticle.getTitle());
         String strDate = dateFormat.format(TopStoriesArticle.getPublished_date());
         this.textViewDate.setText(strDate);
@@ -70,7 +78,9 @@ public class TopStoriesArticleViewHolder extends RecyclerView.ViewHolder impleme
         String temp = TopStoriesArticle.getSection() + ending;
         this.textViewSection.setText(temp);
         this.textViewSection.setTypeface(null, Typeface.BOLD);
-        Glide.with(context).load(TopStoriesArticle.getMultimedia().get(0).getUrl()).apply(options).into(imageView);
+        Picasso.get().load(TopStoriesArticle.getMultimedia().get(0).getUrl())
+//            .apply(options)
+            .into(imageView);
 
         this.callbackWeakRef = new WeakReference<TopStoriesArticleAdapter.Listener>(callback);
     }
@@ -79,11 +89,13 @@ public class TopStoriesArticleViewHolder extends RecyclerView.ViewHolder impleme
     public void onClick(View view) {
         TopStoriesArticleAdapter.Listener callback = callbackWeakRef.get();
         if (callback != null) {
-            callback.onClickDeleteButton(getAdapterPosition());
+            //callback.onClickDeleteButton(getAdapterPosition());
             //callback.changeArticleColor(getAdapterPosition(), this.textView, context);
+            //callback.addToReadQueue(getAdapterPosition());
         }
-        //textView.setTextColor(ContextCompat.getColor(context, R.color.orange));
+        //this.textView.setTextColor(ContextCompat.getColor(context, R.color.colorPrimary));
         //view.setBackgroundColor(ContextCompat.getColor(context, R.color.orange));
+        //ArticleBeenRead.getInstance().setArticleHasBeenRead(topStoriesArticle.getId());
     }
 
     public TextView getTextView() {
